Use the running tmux server to get @plugins or start a new server
I normally start my tmux server with a script because I don't have a
.tmux.config.

i.e.,

    tmux set -g @plugin 'tmux-plugins/tmux-resurrect' \; ...

This change makes it so that @plugins is sourced from the currently
running tmux server instead of spinning up a new one. This also means
that we no longer need to manually parse the configs.

There was also a problem that _tpm_path would not work correctly if
using $XDG_CONFIG_HOME/tmux/tmux.conf because tmux by default doesn't
check that path.

I tested this with both inside and outside tmux by running

    ./tpm/bin/install_plugin